User refactor WIP
