--- app/Http/Controllers/API/Swagger/LogoutSwagger.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\API\Swagger;
-
-use App\Http\Requests\API\LogoutRequest;
-use Illuminate\Http\JsonResponse;
-
-interface LogoutSwagger
-{
-    /**
-     * @OA\Post(
-     *      path="/api/auth/logout",
-     *      description="User logout",
-     *      tags={"Auth"},
-     *      security={
-     *          {"passport": {}},
-     *      },
-     *      @OA\Response(
-     *          response=200,
-     *          description="successful operation",
-     *          @OA\MediaType(
-     *              mediaType="application/json",
-     *          )
-     *      )
-     *   )
-     */
-    public function logout(LogoutRequest $request): JsonResponse;
-}
