+++ app/Services/AuthService.php
@@ -4,7 +4,7 @@ namespace App\Services;
 
 use App\Dto\UserSaveDto;
 use EscolaLms\Core\Enums\UserRole;
-use App\Events\PasswordForgotten;
+use EscolaLms\Auth\Events\PasswordForgotten;
 use App\Repositories\Contracts\UserRepositoryContract;
 use App\Services\Contracts\AuthServiceContract;
 use App\Services\Contracts\UserServiceContract;
